fix: PathResponse distance 필드 double에서 int로 변경

+++ src/main/java/nextstep/subway/applicaion/dto/PathResponse.java
@@ -8,9 +8,9 @@ import java.util.stream.Collectors;
 public class PathResponse {
 
     private final List<StationResponse> stations;
-    private final double distance;
+    private final int distance;
 
-    public PathResponse(List<Station> stations, double distance) {
+    public PathResponse(List<Station> stations, int distance) {
         this.stations = stations.stream().map(StationResponse::from).collect(Collectors.toList());
         this.distance = distance;
     }
@@ -19,7 +19,7 @@ public class PathResponse {
         return stations;
     }
 
-    public double getDistance() {
+    public int getDistance() {
         return distance;
     }
 }
